refactor: modify directory structures regarding creation scripts
