+++ create/deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "start": "deno run -A main.ts"
+    "start": "deno run -A src/main.ts"
   },
   "importMap": "./vendor/import_map.json",
   "lint": {
