+++ create/src/components/createGoogleMeetMeeting.ts
@@ -1,6 +1,6 @@
-import { Page } from "../../mod.ts";
-import waitFor from "../../utils/waitFor.ts";
 import clickGoogleCalendarAddonSaveButton from "../components/clickGoogleCalendarAddonSaveButton.ts";
+import { Page } from "../mod.ts";
+import waitFor from "../utils/waitFor.ts";
 
 const GOOGLE_MEET_BUTTON = "#xAddRtcSel";
