+++ create/src/components/createZoomMeetingUsingChromeExtension.ts
@@ -1,5 +1,5 @@
-import { Page } from "../../mod.ts";
-import waitFor from "../../utils/waitFor.ts";
+import { Page } from "../mod.ts";
+import waitFor from "../utils/waitFor.ts";
 
 const SAVE_BUTTON = "div.BTotkb.JaKw1 > div:nth-child(4) > div";
